requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
pytest>=8.0.0
